Moving the node version check back to the index.
